web-backend/pyproject.toml
[tool.pylint]
output-format = "text"
max-line-length = 100
disable = ["C0111", "C0114", "W0611"]  # Missing docstring, Unused import
ignore = ["migrations", "tests", ".venv", "venv", "docs", "pyproject.toml"]
max-args = 5
allow-wildcard-with-all = false

[tool.ruff]
include = ["src", "tests"]
exclude = ["migrations", ".venv", "venv", "*.pyc"]
line-length = 88
lint.select = [
    "E",  # All errors
    "F",  # All flake8 errors
    "W",  # Warnings
]
lint.ignore = [
    "E501",  # Line too long
]

[tool.isort]
profile = "django"
line_length = 100
skip = ["migrations", "tests", ".venv", "venv", "docs"]

[tool.pytest.ini_options]
asyncio_default_fixture_loop_scope = "function"
# Specify the test paths
testpaths = ["tests"]

# Add any additional options you want to configure
addopts = "-v --tb=short"

# Specify markers for your tests
markers = [
    "slow: mark test as slow"
]

# Configure other pytest options as needed
log_cli_level = "INFO"  # Set the logging level for the command line


[project]
name = "web-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.14.0",
    "asyncpg>=0.30.0",
    "fastapi>=0.115.5",
    "greenlet>=3.1.1",
    "loguru>=0.7.3",
    "mimesis>=18.0.0",
    "pytest-asyncio>=0.25.3",
    "pytest>=8.3.4",
    "python-dotenv>=1.0.1",
    "python-jose>=3.3.0",
    "python-keycloak>=4.7.3",
    "python-multipart>=0.0.19",
    "ruff>=0.7.4",
    "setuptools>=75.8.0",
    "sqlalchemy>=2.0.36",
    "uvicorn>=0.32.0",
    "trio>=0.28.0",
    "aiokafka>=0.12.0",
    "pre-commit>=4.1.0",
]

[dependency-groups]
dev = [
    "isort>=5.13.2",
    "pylint>=3.3.3",
    "ruff>=0.7.4",
]
